Proposal: use test suite structure
