+++ latest/examples/axes/custom_type_axes.json
