+++ latest/examples/image_strict/image.json
