+++ latest/examples/image_strict/image_metadata.json
