+++ latest/examples/image_strict/image_omero.json
