--- latest/examples/invalid/empty_transformations.json
@@ -1,25 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": []
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
