--- latest/examples/invalid/invalid_channels_color.json
@@ -1,48 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [1, 1],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ],
-    "omero": {
-        "channels": [
-            {
-                "active": true,
-                "coefficient": 1.0,
-                "color": 255,
-                "family": "linear",
-                "label": "1234",
-                "window": {
-                    "end": 1765.0,
-                    "max": 2555.0,
-                    "min": 5.0,
-                    "start": 0.0
-                }
-            }
-        ]
-    }
-}
-
