--- latest/examples/invalid/invalid_multiscales_transformations.json
@@ -1,39 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                1,
-                                1
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "coordinateTransformations": [
-                {
-                    "scale": ["invalid"],
-                    "type": "scale"
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
