--- latest/examples/invalid/invalid_transformation_type.json
@@ -1,33 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                1,
-                                1
-                            ],
-                            "type": "translation"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
