--- latest/examples/invalid/missing_axes_name.json
@@ -1,31 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "type": "space",
-                    "units": "micron"
-                },
-                {
-                    "type": "space",
-                    "units": "micron"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                0.13,
-                                0.13
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
