--- latest/examples/invalid/missing_transformations.json
@@ -1,24 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0"
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
