--- latest/examples/invalid/no_axes.json
@@ -1,23 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                1,
-                                1
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
-
