--- latest/examples/invalid/no_datasets.json
@@ -1,20 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "y",
-                    "type": "space",
-                    "units": "micrometer"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [],
-            "version": "0.5-dev"
-        }
-    ]
-}
