--- latest/examples/invalid/no_multiscales.json
@@ -1,4 +1,0 @@
-{
-    "@type": "ngff:Image",
-    "multiscales": []
-}
