--- latest/examples/invalid/too_many_axes.json
@@ -1,51 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "angle",
-                    "type": "custom"
-                },
-                {
-                    "name": "t",
-                    "type": "time"
-                },
-                {
-                    "name": "c",
-                    "type": "channel"
-                },
-                {
-                    "name": "z",
-                    "type": "space"
-                },
-                {
-                    "name": "y",
-                    "type": "space"
-                },
-                {
-                    "name": "x",
-                    "type": "space"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                1,
-                                1,
-                                1,
-                                1,
-                                1,
-                                1
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
