--- latest/examples/invalid/too_many_space_axes.json
@@ -1,41 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "X",
-                    "type": "space"
-                },
-                {
-                    "name": "z",
-                    "type": "space"
-                },
-                {
-                    "name": "y",
-                    "type": "space"
-                },
-                {
-                    "name": "x",
-                    "type": "space"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                1,
-                                1,
-                                1,
-                                1
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
