--- latest/examples/invalid_but_dont_fail/duplicate_axes_name.json
@@ -1,33 +1,0 @@
-{
-    "multiscales": [
-        {
-            "axes": [
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "meter"
-                },
-                {
-                    "name": "x",
-                    "type": "space",
-                    "units": "micrometer"
-                }
-            ],
-            "datasets": [
-                {
-                    "path": "0",
-                    "coordinateTransformations": [
-                        {
-                            "scale": [
-                                0.13,
-                                0.13
-                            ],
-                            "type": "scale"
-                        }
-                    ]
-                }
-            ],
-            "version": "0.5-dev"
-        }
-    ]
-}
