+++ latest/examples/multiscales_strict/multiscales_example.json
