+++ latest/examples/multiscales_strict/multiscales_transformations.json
