+++ latest/tests/test_validation.py
@@ -1,73 +1,49 @@
+"""
+This test consumes https://github.com/json-schema-org/JSON-Schema-Test-Suite#structure-of-a-test
+styled JSON tests.
+"""
 import json
 import glob
 
 import pytest
 
-from jsonschema import RefResolver, Draft202012Validator
+from jsonschema import RefResolver, Draft7Validator
 from jsonschema.exceptions import ValidationError
 
 
-with open('schemas/image.schema') as f:
-    image_schema = json.load(f)
-with open('schemas/strict_image.schema') as f:
-    strict_image_schema = json.load(f)
-schema_store = {
-    image_schema['$id']: image_schema,
-    strict_image_schema['$id']: strict_image_schema,
-}
-
-resolver = RefResolver.from_schema(image_schema, store=schema_store)
-validator = Draft202012Validator(image_schema, resolver=resolver)
-strict_validator = Draft202012Validator(strict_image_schema, resolver=resolver)
-
-valid_strict_files = list(glob.glob("examples/valid_strict/*.json"))
-valid_files = list(glob.glob("examples/valid/*.json"))
-invalid_files = list(glob.glob("examples/invalid/*.json"))
-invalid_but_dont_fail_files = list(
-    glob.glob("examples/invalid_but_dont_fail/*.json"))
+def pytest_generate_tests(metafunc):
+    if "suite" in metafunc.fixturenames:
+        suites = []
+        for filename in glob.glob("tests/*.json"):
+            with open(filename) as o:
+                data = json.load(o)
+            schema = data["schema"]
+            for test in data["tests"]:
+                suites.append((schema, test))
+        metafunc.parametrize("suite", suites, indirect=True)
 
 
-def ids(files):
-    return [str(x).split("/")[-1][0:-5] for x in files]
+@pytest.fixture
+def suite(request):
+    return request.param
 
 
-@pytest.mark.parametrize(
-    "testfile", valid_strict_files, ids=ids(valid_strict_files))
-def test_valid_strict(testfile):
-    with open(testfile) as f:
-        data = ''.join(line for line in f if not line.lstrip().startswith('//'))
-        jsondata = json.loads(data)
-        validator.validate(jsondata)
-        strict_validator.validate(jsondata)
+def test_run(suite):
+    schema, test = suite
 
+    # Load schema
+    # Currently missing the "id" field
+    with open(schema["id"]) as f:
+        schema = json.load(f)
+    schema_store = {
+        schema['$id']: schema
+    }
+    resolver = RefResolver.from_schema(schema, store=schema_store)
+    validator = Draft7Validator(schema, resolver=resolver)
 
-@pytest.mark.parametrize("testfile", valid_files, ids=ids(valid_files))
-def test_valid_files(testfile):
-    with open(testfile) as f:
-        json_file = json.load(f)
-        validator.validate(json_file)
+    # Load data and test
+    jsondata = test["data"]
+    valid = test["valid"]
+    if not valid:
         with pytest.raises(ValidationError):
-            strict_validator.validate(json_file)
-
-
-@pytest.mark.parametrize("testfile", invalid_files, ids=ids(invalid_files))
-def test_invalid(testfile):
-    with open(testfile) as f:
-        json_file = json.load(f)
-        with pytest.raises(ValidationError):
-            validator.validate(json_file)
-        with pytest.raises(ValidationError):
-            strict_validator.validate(json_file)
-
-
-@pytest.mark.xfail
-@pytest.mark.parametrize(
-    "testfile", invalid_but_dont_fail_files,
-    ids=ids(invalid_but_dont_fail_files))
-def test_invalid_but_dontfail(testfile):
-    with open(testfile) as f:
-        json_file = json.load(f)
-        with pytest.raises(ValidationError):
-            validator.validate(json_file)
-        with pytest.raises(ValidationError):
-            strict_validator.validate(json_file)
+            validator.validate(jsondata)
